requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==3.2.20
djangorestframework==3.14.0
mysqlclient==2.1.1
packaging==23.1
pytz==2023.3
querystring-parser==1.2.4
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
